+++ CHANGELOG.md
@@ -6,6 +6,10 @@ patch-level version changes can be found in [commit messages](../../commits/mast
 - Hide binary in sound and image files (switched from text to binary for this)
 - Can now hide whole images inside other images!
 - Use map and otp lsb encodings for sound
+- Image mode is preserved
+- Removed CLI
+- Can hide data in Microsoft Office files
+- Can hide data in OpenOffice/ LibreOffice files
 
 ## 2020 - 2020/05/22
 - First release